Add Judge0 API flag

+++ utils/delete-post.ts
@@ -1,6 +1,7 @@
 import Router from "next/router";
 
 async function deletePost(id: number): Promise<void> {
+  if (isNaN(id)) throw new Error("Invalid post ID.");
   await fetch(`/api/c/${id}`, {
     method: "DELETE",
   });
